Handle boolean attributes with :attrs
When passing a boolean attribute to the :attrs modifier, any true value will be output
as just the key, without any value. When the value is false, then the attribute will not
show at all.

This behavior relies on the value being a real boolean, not just a truthy value.